Fixing warnings for -Wall in the io, linalg and mesh folders
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@872 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ mesh/src/common/ConformingTetrahedralMesh.cpp
@@ -74,7 +74,7 @@ void ConformingTetrahedralMesh<ELEMENT_DIM, SPACE_DIM>::ConstructFromMeshReader(
 		    
 			for (int j=0; j < nodes_size; j++)
 			{
-				assert(node_indices[j] < temp_nodes.size());
+				assert(node_indices[j] < (int) temp_nodes.size());
 				nodes.push_back(&temp_nodes[node_indices[j]]);
 			}
 						
@@ -339,7 +339,7 @@ void ConformingTetrahedralMesh<ELEMENT_DIM, SPACE_DIM>::AddNode(Node<SPACE_DIM> 
 template<int ELEMENT_DIM, int SPACE_DIM>
 Node<SPACE_DIM> *ConformingTetrahedralMesh<ELEMENT_DIM, SPACE_DIM>::GetNodeAt(long index)
 {
-	assert(index < mNodes.size());
+	assert(index < (long) mNodes.size());
     return &(mNodes[index]);
 }
